Fix: Incorrect type of GetMapSetId in OsuMemory sig reader

+++ OsuMemoryDataProvider/IOsuMemoryReader.cs
@@ -22,7 +22,7 @@ namespace OsuMemoryDataProvider
         float GetMapCs();
         float GetMapHp();
         float GetMapOd();
-        float GetMapSetId();
+        int GetMapSetId();
         string GetSkinFolderName();
         string GetOsuFileName();
         string GetMapFolderName();
